Cleaning_FirstDraft
Finalisation of the first pass of cleaning, all edits to code made in Data_Cleaniing.R.

Extra steps need to be taken to cross-check the data against VCOSS's pre-prepared data, the first of which will be consultation.

This project turned collaborative today, so branches will start to be created/used/merged.